+++ backend/src/main/java/com/tatanka/dsvendas/entities/Sale.java
@@ -23,7 +23,7 @@ public class Sale {
 	
 	private Integer deals;
 	
-	private Integer amount;
+	private Double amount;
 	
 	private LocalDate date;
 	
@@ -33,7 +33,7 @@ public class Sale {
 	
 	public Sale() {}
 
-	public Sale(Long id, Integer visited, Integer deals, Integer amount, LocalDate date, Seller seller) {
+	public Sale(Long id, Integer visited, Integer deals, Double amount, LocalDate date, Seller seller) {
 		super();
 		this.id = id;
 		this.visited = visited;
@@ -67,11 +67,11 @@ public class Sale {
 		this.deals = deals;
 	}
 
-	public Integer getAmount() {
+	public Double getAmount() {
 		return amount;
 	}
 
-	public void setAmount(Integer amount) {
+	public void setAmount(Double amount) {
 		this.amount = amount;
 	}
